Refactor `toNotificationEvents`
