--- packages/server/src/notifier/model/notificationEvents/fromEvents/index.ts
@@ -1,1 +1,0 @@
-export * from './forum'
